Add new_span and update_span configs to logger
The intention of this change is simple: letting
the user of the library configure the span
creation and update it after the request.
This would be interesting, because it'd let
the user change the field names and even add
more fields without disrupting the lib.

The Default trait allows for people that do not
care about this customization to leverage the sane
defaults as it were before, so I think it's win win.