feat: add `http.peer_ip` to the root span
`realip_remote_addr` can be spoofed without using proxy or not
configuring it properly. `peer_addr` is considered more trustworthy for
security-sensitive operations, which is the actual socket address of the
client when not behind any proxy, and the proxy's address otherwise.

When we are using a trusted and properly configured proxy, `peer_addr`
should return the proxy's address, and `realip_remote_addr` should return
a trusted client address.